optical_sim/requirements.txt
numpy>=1.24
pyyaml>=6.0
matplotlib>=3.7
# Optional — uncomment when integrating:
waveprop
laserbeamsize
